chatbot api pydantic 스키마 파일 분리
